+++ doc/Changelog.md
@@ -3,6 +3,8 @@ Changelog {#Changelog}
 
 # Release 0.4 (git master) {#Release040}
 
+* [#86](https://github.com/BlueBrain/Livre/pull/86):
+  Removed obsolete limitation to open only regular power-of-two volumes
 * [#84](https://github.com/BlueBrain/Livre/pull/84):
   Cache statistics are printed when showing statistics (press 's' key)
 * [#84](https://github.com/BlueBrain/Livre/pull/84):